refactor(frontend/auth): extract redirectToLogin into domain module
`redirectToLogin` was buried inside graphqlClientApollo.tsx, which meant
the REST middleware (and any future caller) had to either duplicate the
"clear tokens + hard-nav with ?from" logic or import across a layer it
shouldn't depend on.

Move the helper + its `__navigation` test holder to
entities/authentication/domain/redirect-to-login.ts. No behaviour change;
Apollo and its test now import from the new location.
