+++ frontend/app/src/shared/api/graphql/graphqlClientApollo.tsx
@@ -17,8 +17,8 @@ import { ALERT_TYPES, Alert } from "@/shared/components/ui/alert";
 import { CONFIG } from "@/shared/config/config";
 
 import { ACCESS_TOKEN_KEY } from "@/entities/authentication/constants";
+import { redirectToLogin } from "@/entities/authentication/domain/redirect-to-login";
 import { refreshAccessTokenQueryOptions } from "@/entities/authentication/ui/queries/refresh-access-token.query";
-import { removeTokensInLocalStorage } from "@/entities/authentication/utils";
 
 export const defaultOptions: DefaultOptions = {
   watchQuery: {
@@ -184,30 +184,6 @@ function retryWithRefreshedToken(
         observer.error(err);
       });
   });
-}
-
-// Helper: token is invalid or missing — clear local credentials and bounce
-// to /login. Hard-navigates because errorLink runs outside React Router,
-// and the AuthProvider's localStorage hook does not re-render on external
-// writes. Skips the redirect if we're already on /login to avoid loops.
-//
-// Encodes the current path as `?from=…` so `LoginPage` can route the user
-// back to where they were after re-authenticating. The hard nav means
-// `location.state` is gone, so the query string is the only carrier left.
-// Holder so tests can stub the hard-nav without touching `window.location`
-// (which is non-configurable in real browsers — vitest's browser mode hits
-// that wall the moment you try to spy on `assign`). Production reads
-// through this same reference, so the indirection costs one property lookup.
-export const __navigation = {
-  assign: (url: string) => window.location.assign(url),
-};
-
-function redirectToLogin(): void {
-  removeTokensInLocalStorage();
-  if (window.location.pathname === "/login") return;
-
-  const from = window.location.pathname + window.location.search + window.location.hash;
-  __navigation.assign(`/login?from=${encodeURIComponent(from)}`);
 }
 
 // Helper: surface an error to the user. Calls operation.context's
